requirement.txt
dnspython==2.0.0
Flask==1.1.1
Flask-PyMongo==2.3.0
gunicorn==20.0.4
pymongo==3.11.0
python-dotenv==0.14.0